Split translation and verification features.
